requirements.txt
requests==2.18.4
bs4==0.0.1
alauda-pytest==5.2.3
pytest-html==2.0.0
pytest-rerunfailures==8.0
flake8==3.7.9
pyyaml==5.3.1
python-ldap==3.1.0
websockets==8.1
asyncio==3.4.3
jinja2==2.11.1
alauda-xdist==1.30.1
pytest-ordering==0.6
dpath==2.0.1
flask==1.1.1
xlrd==1.1.0
IPy==1.0
TestLink-API-Python-client==0.8.1
